Sets up a NODE_ENV=local flag
Previously I was looking for `tessel.port` but I can't remember why that
was a thing and it doesn't seem to be working. I am wondering if maybe the
tessel is plugged in but the IR module was not connected? I am not using
my board right now so I am leaving the line in.